requirements.txt
matplotlib==3.5.1
numpy>=1.20.3
opencv-python-headless==4.5.3.56
pandas==1.5.3
Pillow==10.1.0
scikit-image==0.20.0
scikit-learn==1.0.2
streamlit==1.28.2
tensorflow==2.13.0
huggingface-hub==0.15.1
timm==0.9.12
scikit-image==0.20.0
ultralytics==8.0.235
